hirelens-backend/requirements.txt
# Flask & Web Framework
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
flask-migrate
werkzeug

# Database
pymongo
dnspython

# Authentication & OAuth
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
oauthlib
requests-oauthlib

# Machine Learning & AI
torch
torchvision
transformers
accelerate
mediapipe

# Computer Vision
opencv-python
numpy
pandas

# Audio Processing & Speech Recognition
SpeechRecognition
pyaudio
pydub
librosa
ffmpeg-python
RealtimeSTT
pocketsphinx

# Utilities
python-dotenv
requests
cachetools

# Network & HTTP
certifi
charset-normalizer
httplib2
idna
urllib3
